Implement client_ip matcher, deprecate remote_ip's forwarded option
